+++ CHANGELOG.md
@@ -1875,3 +1875,4 @@
 - [2026-06-10 13:58] test(storage): add unit tests for storage business logic
 - [2026-06-10 14:24] test(test): implement widget integration tests for test
 - [2026-06-10 14:42] fix(state): resolve race condition during state initialization
+- [2026-06-10 14:54] test(analytics): implement widget integration tests for analytics